+++ sdk/package/lib/StartSdk.ts
@@ -11,6 +11,7 @@ import * as patterns from '../../base/lib/util/patterns'
 import { Backups } from './backup/Backups'
 import {
   smtpInputSpec,
+  smtpShape,
   systemSmtpSpec,
   customSmtp,
   smtpProviderVariants,
@@ -408,6 +409,7 @@ export class StartSdk<Manifest extends T.SDKManifest> {
       },
       inputSpecConstants: {
         smtpInputSpec,
+        smtpShape,
         systemSmtpSpec,
         customSmtp,
         smtpProviderVariants,